ServerModule, BouncerModule -> Disfigure.Server
I've moved both the Server and Bouncer modules to their own project & namespace. This is an effort to separate the generic (Module) base class from the concrete implementation for my specific packet type and protocol.

If, for instance, one wanted to create an IRC implementation of the `Module` class, they would have to implement their own custom `ServerModule` and `IRCPacket` types (names improvised).

This is overall not so much a *change*, as a shift in architecture. My custom implementations of a server and bouncer module are now effectively separate from the base `Disfigure` library, so must be explicitly imported to be used.